Initial changes to fix rect timing.
This is an initial attempt at resolving a potential issue wherein
updates would be sent to the server delayed relative to the framebuffer
sync they were associated with. This causes the server to receive
DISPLAY_UPDATE events that do not correspond to the framebuffer in the
shared memory when the message is received, which leads to graphics
corruption.

This commit eliminates the buffer update queue, and instead accumulates
changes in the QEMU framebuffer and changes corresponding to the current
"version" of the shared memory buffer separately. The two states are
synced along with the framebuffer in the mux_display_refresh callback.

In this way, this patch hopes to avoid consistency and timing issues
regarding the change regions of the framebuffer.